BUG/MINOR: quic: Missing succesful SSL handshake backend trace (OpenSSL 3.5)
This very minor bug impacts only the backend when compiled against OpenSSL 3.5
with QUIC API (HAVE_OPENSSL_QUIC).

The "SSL handshake OK" trace was not dumped.
Modify the concerned code section which is a bit ugly and simplify. The TRACE()
call is done at a unique location for now on.

Should be backported to 3.2 to ease any further backport.